feat(Calendar) remove context WIP
