--- web-client/src/actions/CalendarAction.ts
@@ -1,5 +1,0 @@
-import { EventFormated } from '../interfaces/eventInterface';
-
-export type CalendarAction =
-  | { type: 'CALENDAR_ADD_EVENTS'; calendarEvents: EventFormated[] }
-  | { type: 'CALENDAR_EMPTY'; calendarEvents: [] };
