+++ web-client/src/containers/calendar-ctnr/CalendarCtnr.tsx
@@ -1,45 +1,14 @@
-import React, { useEffect, useState, useContext } from 'react';
+import React from 'react';
 import FullCalendar, { EventDropArg } from '@fullcalendar/react';
 import dayGridPlugin from '@fullcalendar/daygrid';
 import interactionPlugin from '@fullcalendar/interaction';
-import { useQuery, useMutation } from '@apollo/client';
+import { useMutation } from '@apollo/client';
 
-import UserContext from '../../contexts/UserContext';
-import CalendarContext from '../../contexts/CalendarContext';
-import { GET_EVENTS_CREATED_BY_USER_AUTHENTICATED } from '../../queries/userQueries';
+import { useDisplayCalendarEvents } from '../../hooks/calendarHooks';
 import { UPDATE_EVENT_BY_ID } from '../../queries/calendarEventQueries';
-import {
-  EventFromDataInterface,
-  EventFormated,
-} from '../../interfaces/eventInterface';
-
-const getEventsFormatted = (
-  data: [EventFromDataInterface]
-): EventFormated[] => {
-  const formatedEventsArray = data.map((event) => {
-    return {
-      id: event.id,
-      title: event.eventTitle,
-      start: event.eventStart,
-      end: event.eventEnd,
-      allDay: event.eventAllDay,
-      eventContent: event.eventContent,
-    };
-  });
-  return formatedEventsArray;
-};
 
 const CalendarCtnr = (): JSX.Element => {
-  const userLoggedIn = useContext(UserContext);
-  const calendarContext = useContext(CalendarContext);
-  const [eventsArrayFormated, setEventsArrayFormated] = useState<
-    EventFormated[]
-  >([]);
-
-  const { data } = useQuery(GET_EVENTS_CREATED_BY_USER_AUTHENTICATED, {
-    variables: { id: userLoggedIn.state.userLoggedInDetails?.id },
-  });
-
+  const { eventsArrayFormated, loading } = useDisplayCalendarEvents();
   const [updateCalendarEvent] = useMutation(UPDATE_EVENT_BY_ID);
 
   const eventDatasOnDragStop = async (eventInfo: EventDropArg) => {
@@ -54,32 +23,27 @@ const CalendarCtnr = (): JSX.Element => {
     });
   };
 
-  useEffect(() => {
-    if (data) {
-      setEventsArrayFormated(getEventsFormatted(data.user.eventsCreatedByUser));
-      calendarContext.calendarDispatch({
-        type: 'CALENDAR_ADD_EVENTS',
-        calendarEvents: getEventsFormatted(data.user.eventsCreatedByUser),
-      });
-    }
-  }, [data]);
   return (
     <section className="calendar-section">
-      <FullCalendar
-        plugins={[dayGridPlugin, interactionPlugin]}
-        locale="fr"
-        headerToolbar={{
-          left: 'prev,next',
-          center: 'title',
-          right: 'dayGridMonth,dayGridWeek,dayGridDay',
-        }}
-        editable={true}
-        droppable={true}
-        initialView="dayGridMonth"
-        weekends={false}
-        events={eventsArrayFormated}
-        eventDrop={eventDatasOnDragStop}
-      />
+      {loading ? (
+        <p>Loading...</p>
+      ) : (
+        <FullCalendar
+          plugins={[dayGridPlugin, interactionPlugin]}
+          locale="fr"
+          headerToolbar={{
+            left: 'prev,next',
+            center: 'title',
+            right: 'dayGridMonth,dayGridWeek,dayGridDay',
+          }}
+          editable={true}
+          droppable={true}
+          initialView="dayGridMonth"
+          weekends={false}
+          events={eventsArrayFormated}
+          eventDrop={eventDatasOnDragStop}
+        />
+      )}
     </section>
   );
 };
