+++ web-client/src/style.css
@@ -121,7 +121,7 @@
   position: relative;
   box-sizing: border-box;
   margin: 20px 0;
-  background-color: #49BEB7;
+  background-color: #49beb7;
   border-radius: 5px;
   width: 50%;
   display: flex;
@@ -138,11 +138,11 @@
   box-sizing: border-box;
   width: 100%;
   padding: 0 35px;
-  background-color: #49BEB7;
+  background-color: #49beb7;
   border: none;
   cursor: pointer;
   font-size: 1rem;
-  color: #F6F5F5;
+  color: #f6f5f5;
   max-height: 0;
   transition: max-height 250ms ease-in-out 250ms, padding 250ms ease-in-out 250ms;
 }
@@ -191,7 +191,7 @@
     position: fixed;
     top: 0;
     left: 0;
-    width: 300px;
+    width: 80%;
     transform: translateX(0%);
     transition: transform 350ms ease-in-out;
   }
@@ -211,7 +211,7 @@
 }
 @media only screen and (max-width: 760px) {
   .left-aside.left-aside-visible {
-    width: 300px;
+    width: 80%;
     transform: translateX(-100%);
   }
 }
@@ -252,10 +252,8 @@
     max-height: 100vh;
     position: absolute;
     top: 0;
-    left: 100%;
-    width: 80%;
-    transform: translateX(-100%);
-    transition: transform 350ms ease-in-out;
+    right: 0;
+    width: 0;
   }
 }
 .right-aside > * + * {
@@ -276,8 +274,7 @@
 }
 @media only screen and (max-width: 760px) {
   .right-aside.ra-visible {
-    max-width: 300px;
-    transform: translateX(0%);
+    width: 80%;
   }
 }
 .right-aside .ra-title {
@@ -317,6 +314,16 @@
 
 .channels-ctnr {
   width: 100%;
+}
+
+/* ************************** */
+/* ************************** */
+/* ***** CALENDAR CTNR ****** */
+/* ************************** */
+/* ************************** */
+.calendar-section {
+  margin-top: 25px;
+  width: 80%;
 }
 
 .channel-messages {
@@ -423,16 +430,6 @@
 }
 .channel-messages .cm-content-wrapper .cmcw-form .cmcwf-footer .cmcwff-send-button .cmcwff-send-icon {
   color: #f6f5f5;
-}
-
-/* ************************** */
-/* ************************** */
-/* ***** CALENDAR CTNR ****** */
-/* ************************** */
-/* ************************** */
-.calendar-section {
-  margin-top: 25px;
-  width: 80%;
 }
 
 /* ************************** */
@@ -825,7 +822,7 @@
 .calendar-form {
   box-sizing: border-box;
   min-width: 320px;
-  background-color: #49BEB7;
+  background-color: #49beb7;
   padding: 0 30px;
   width: 100%;
   display: flex;
@@ -880,6 +877,67 @@
 }
 .calendar-form .cfic-submit:hover {
   background-color: #cbf3f0;
+}
+
+.message {
+  display: flex;
+  flex-direction: column;
+}
+.message .m-header {
+  margin: 1.5rem 0 0.5rem 0;
+}
+.message .m-header.mh-on-left {
+  align-self: flex-start;
+}
+.message .m-header.mh-on-right {
+  align-self: flex-end;
+}
+.message .m-content-wrapper {
+  width: 100%;
+  display: flex;
+  align-items: flex-end;
+  margin: 0.1rem 0;
+}
+.message .m-content-wrapper.on-left {
+  flex-direction: row-reverse;
+  justify-content: flex-end;
+}
+.message .m-content-wrapper.on-left > * + * {
+  margin-right: 0.5rem;
+}
+.message .m-content-wrapper.on-right {
+  justify-content: flex-end;
+}
+.message .m-content-wrapper.on-right > * + * {
+  margin-left: 0.5rem;
+}
+.message .m-content {
+  border-radius: 5px;
+  padding: 0.5em;
+  width: max-content;
+  max-width: 75%;
+  word-break: break-word;
+}
+.message .m-content.color-left {
+  background-color: rgba(136, 225, 220, 0.5);
+}
+.message .m-content.color-right {
+  background-color: rgba(16, 176, 168, 0.5);
+}
+.message .m-time {
+  font-size: 0.8rem;
+  font-style: italic;
+}
+
+.textarea {
+  resize: none;
+  padding: 0.7rem 0.5rem;
+  background-color: #f6f5f5;
+  border: none;
+  outline: none;
+}
+.textarea::placeholder {
+  font-size: 0.8rem;
 }
 
 * {
